+++ tests/testthat/test-parameterSweep.R
@@ -50,6 +50,12 @@ test_that("Support functions and diagnostics", {
 
   diagnostic <- computeTimeStability(studyPop, model)
   expect_is(diagnostic, "data.frame")
+
+  mdrr <- computeMdrr(sccsIntervalData = sccsIntervalData, exposureCovariateId = 1000)
+  expect_lt(mdrr$mdrr, Inf)
+
+  mdrr <- computeMdrr(sccsIntervalData = sccsIntervalData, exposureCovariateId = 9999)
+  expect_equal(mdrr$mdrr, Inf)
 })
 
 test_that("Parameter sweep", {